Rework Gamepad driver to add a deadzone
This reworking of the gamepad driver moves the driver configuration
into its own type and object, which is now passed in through to
configure the driver. This also moves around the driver function to
break it into smaller parts.

Notably this change introduces a deadzone configuration argument to
allow the driver to apply a deadzone to the gamepad:

    run(App, {
        gamepad: makeGamepadDriver({
            deadzone: DeadzoneImpl(implConfig),
        }),
    });

The current DeadzoneFunction type is a simple mapping of an axis
value to a new value, which works fine for the AxialDeadzone
implemented here, but may not work when the deadzone implementation
needs access to multiple axes. We can update that when it's needed.

Proxy usage
-----------

Admittedly, I'm surprised that this seems to not have any noticable
impact on performance. I've proxied all of the gamepad axes through
to be able to apply the deadzone function to their values so we end
up with a call chain like so:

    Gamepad
    ` gamepad.buttons
     ` proxy
      ` buttons[0]
       ` proxy
        ` GamepadButton
         ` proxy
          ` button.value

And like so for axes:

    Gamepad
    ` gamepad.axes
     ` proxy
      ` axes[0]
       ` proxy
        ` value

We will need to watch this to see how it performs.